Offline files verbeterd
Het is nu mogelijk om offline files met funda HTML te uploaden naar het script waarna deze de informatie verwerkt in de database.

+++ check_offline.php
@@ -1,69 +1,139 @@
 <?php
 include_once(__DIR__.'/include/config.php');
-include_once('include/HTML_TopBottom.php');
+include_once(__DIR__ .'/include/HTML_TopBottom.php');
 include_once($cfgGeneralIncludeDirectory.'class.phpPushover.php');
+$minUserLevel = 3;
+$cfgProgDir = 'auth/';
+include($cfgProgDir. "secure.php");
+
 $db = connect_db();
 
-for($i=0; $i<2 ; $i++) {
-	$String = $block = array();
+$String = $block = array();
 	
-	if($i == 0) {
-		$pageDir = $offlineDir.'koop/';
-		$verkocht = false;
-	} else {
-		$pageDir = $offlineDir.'verkocht/';
-		$verkocht = true;
-	}
-		
-	if ($handle = opendir($pageDir)) {
-		while (false !== ($entry = readdir($handle))) {
-			if ($entry != "." && $entry != "..") {
-				$files[] = $entry;
-			}
-	  }
-		closedir($handle);
-	}
-	
-	$debug = 0;
+if ($handle = opendir($offlineDir)) {
+	while (false !== ($entry = readdir($handle))) {
+		if ($entry != "." && $entry != "..") {
+			$files[] = $entry;
+		}
+	}	
+	closedir($handle);
+}
 
-	# Doorloop alle zoekopdrachten
-	foreach($files as $file) {
-		# Alles initialiseren
-		set_time_limit (30);
-		$bestand = $pageDir.$file;
-		$AdressenArray = array();
-			
-		# Bestandsnaam is opgebouwd volgens
-		#		$debug_filename = 'funda_'. $OpdrachtID .'_'. $p .'.htm';
-		$delen = explode('_', $file);
-		$OpdrachtID = $delen[1];
-		$p = substr($delen[2], 0, -5);
+$opdrachten = getZoekOpdrachten($_SESSION['account'], '', false);
+
+$debug = 0;
+
+# Doorloop alle offline-bestanden
+foreach($files as $file) {
+	# Alles initialiseren
+	set_time_limit (30);	
+	$String = $AdressenArray = array();
+	$OpdrachtID = 0;
+	$detail = $verkocht = $overzicht = false;
+	
+	$bestand = $offlineDir.$file;
 		
+	$fp = fopen($bestand, 'r+');
+	$contents = fread($fp, filesize($bestand));
+	fclose($fp);
+	
+	# Achterhalen wat voor een soort pagina het is
+	# 	Overzichtspagina met te koop staande huizen
+	# 	Overzichtspagina met verkochte huizen
+	# 	Detail pagina van een individueel huis	
+	$appHeaderLink	= getString('appheader-English-link" class="app-header__link" href="', '" hreflang=', $contents, 0);
+	$pageURL				= getString('/koop/', '', $appHeaderLink[0], 0);	
+	$zoekURL				= '/koop/'.$pageURL[0];
+	$cleanZoekString = str_replace('/verkocht/', '/', $zoekURL);
+		
+	foreach($opdrachten as $opdracht) {
+		$opdrachtData = getOpdrachtData($opdracht);
+		$cleanOpdrachtURL = str_replace('http://www.funda.nl', '', $opdrachtData['url']);
+		
+		# Zoeken naar tekoopstaande huizen		
+		# /p2/ en /p3/ etc. moet er van worden afgeknipt dus even zo
+		if(substr($zoekURL, 0, strlen($cleanOpdrachtURL)) == $cleanOpdrachtURL){
+			$OpdrachtID = $opdracht;
+			$overzicht = true;
+			$verkocht = false;
+		}
+		
+		# Zoeken naar verkochte huizen
+		if(substr($cleanZoekString, 0, strlen($cleanOpdrachtURL)) == $cleanOpdrachtURL AND strpos($zoekURL, '/verkocht/')){
+			$OpdrachtID = $opdracht;
+			$overzicht = true;
+			$verkocht = true;
+		}		
+	}
+	
+	# Het lijkt geen overzichtspagina te zijn
+	# Onderzoek de optie detailpagina
+	if(!$overzicht) {
+		if(strpos($zoekURL, '/verkocht/')) {
+			$verkocht		= true;
+		}
+		
+		$mappen = explode("/", $zoekURL);
+		 
+		if($verkocht) {
+			$delen 		= explode("-", $mappen[4]);
+		} else {
+			$delen 		= explode("-", $mappen[3]);
+		}
+		
+		$fundaID	= $delen[1];
+		
+		if(is_numeric($fundaID)) {
+			$detail = true;
+		}
+	}	
+	
+	/*
+	if($verkocht) {
+		if($overzicht) {
+			$String[] = $file .' -> VERKOCHT | '. $OpdrachtID .'<br>';
+		} else {
+			$String[] = $file .' -> VERKOCHT | '. $fundaID .'<br>';
+		}
+	} else {
+		if($overzicht) {
+			$String[] = $file .' -> '. $OpdrachtID .'<br>';
+		} else {
+			$String[] = $file .' -> '. $fundaID .'<br>';
+		}
+	}
+	*/
+	
+	# 
+	# De routine als geen van beide gevonden is
+	#
+	if(!$overzicht AND !$detail) {
+		$String[] = "Het type pagina van <a href='$bestand'>$file</a> kan niet worden bepaald<br>";
+	
+	
+	
+	# 
+	# De routine als het een overzichtspagina is
+	#
+	} elseif($overzicht) {
 		$OpdrachtData			= getOpdrachtData($OpdrachtID);
 		$PushMembers			= getMembers4Opdracht($OpdrachtID, 'push');
 		
 		$OpdrachtURL	= $OpdrachtData['url'];
 		
 		if($verkocht) {
-			toLog('info', $OpdrachtID, '', 'Start inladen verkochte huizen voor '. $OpdrachtData['naam'] .", pagina $p");
+			toLog('info', $OpdrachtID, '', 'Inladen pagina verkochte huizen voor '. $OpdrachtData['naam']);
 		} else {
-			toLog('info', $OpdrachtID, '', 'Start inladen huizen voor  '. $OpdrachtData['naam'] .", pagina $p");
+			toLog('info', $OpdrachtID, '', 'Inladen pagina voor '. $OpdrachtData['naam']);
 		}
-		
-		//$String[] = "<a href='$OpdrachtURL'>". $OpdrachtData['naam'] ."</a><br>\n";
-		//$block[] = implode("\n", $String);		
 			
-		$fp = fopen($bestand, 'r+');
-		$contents = fread($fp, filesize($bestand));
-		fclose($fp);
-		
 		# Code opknippen zodat er een array met HTML-code voor een huis ontstaat		
 		$Huizen			= explode('<div class="search-result-media">', $contents);
 		$Huizen			= array_slice($Huizen, 1);		
 		$NrPageHuizen		= count($Huizen);
 		
 		if($debug == 1) {
-			$block[] = "Aantal huizen in <a href='bestand'>pagina $p</a> : ". $NrPageHuizen ."<br>\n";
+			$block[] = "Aantal huizen in <a href='$bestand'>$file</a> : ". $NrPageHuizen ."<br>\n";
 		}
 		
 		# Doorloop nu alle gevonden huizen op de overzichtspagina
@@ -110,7 +180,7 @@ for($i=0; $i<2 ; $i++) {
 				
 				# Aanvinken om in een later stadium de details op te vragen
 				mark4Details($data['id']);
-			}
+			} 
 			
 			# Huis is al bekend bij het script
 			# We moeten dus aangeven dat hij nog steeds op de markt is
@@ -207,19 +277,73 @@ for($i=0; $i<2 ; $i++) {
 		
 				$NewAddress[] = $data['adres'];
 					
-				if($debug == 0 AND !$verkocht)	sendPushoverNewHouse($data['id'], $OpdrachtID);
-			} elseif(changedPrice($data['id'], $data['prijs'], $OpdrachtID))	sendPushoverChangedPrice($data['id'], $OpdrachtID);
+				if($debug == 0 AND !$verkocht) {
+					sendPushoverNewHouse($data['id'], $OpdrachtID);
+				}
+			} elseif(changedPrice($data['id'], $data['prijs'], $OpdrachtID)) {
+				sendPushoverChangedPrice($data['id'], $OpdrachtID);
+			}
 			
-			if(!$verkocht)	addUpdateStreetDb(extractStreetFromAdress($data['adres']), $data['plaats']);
+			if(!$verkocht) {
+				addUpdateStreetDb($data['straat'], $data['plaats']);
+			}
 		}
 	
-		$String[] = "<a href='$bestand'>Pagina $p</a> van <a href='$OpdrachtURL'>". $OpdrachtData['naam'] ."</a> verwerkt ($file) en ". count($AdressenArray)  ." huizen gevonden<br>";
+		$String[] = "<a href='$bestand'>Overzicht</a>". ($verkocht ? ' met verkochte huizen ' : ' ')."van <a href='$OpdrachtURL'>". $OpdrachtData['naam'] ."</a>; ". count($AdressenArray)  ." huizen gevonden<br>";
 			
-		toLog('debug', $OpdrachtID, '', "Einde pagina $p (". count($AdressenArray) ." huizen)");
+		toLog('debug', $OpdrachtID, '', "Einde pagina (". count($AdressenArray) ." huizen)");		
+	
+	
+	
+	
+	# 
+	# De routine als het een detailspagina is
+	#
+	} elseif($detail) {
+		$allData = extractFundaDataFromPage($contents);
+		$data = $allData[0];
+		
+		if($fundaID != $data['id']) {
+			$String[] = "Klopt dit wel ?";
+		} else {
+			$String[] = "Details van <a href='". $ScriptURL ."admin/edit.php?id=$fundaID'>". $data['adres'] ."</a> ingelezen<br>\n";
+			
+			# Als wij een huis niet kennen klopt er iets niet
+			if(!knownHouse($fundaID)) {
+				toLog('error', '', $fundaID, 'Huis niet bekend');
+					
+			# Meestal zal het huis wel bekend zijn
+			} else {
+				updateHouse($data, $allData[1]);
+				addCoordinates($data['adres'], $data['PC_c'], $data['plaats'], $fundaID);
+				updatePrice($fundaID, $data['prijs'], time());
+				
+				# Als hij nog niet verkocht is moeten wij dat aangeven
+				if($data['verkocht'] != 1) {
+					updateAvailability($fundaID);
+					addUpdateStreetDb($data['straat'], $data['plaats']);
+		
+				# Als hij wel verkocht is moeten we de administratie daarvan even bijwerken
+				} else {
+					$temp = updateVerkochtDataFromPage($data, $allData[1]);
+					$String[] = implode("<br>\n", $temp)."<br>\n";
+				}
+				
+				toLog('info', '', $fundaID, 'Offline pagina ingeladen');
+				remove4Details($fundaID);
+			}
+		}
+	}
+	
+	# Alleen als het een overzichts-pagina of detail-pagina is
+	#		(lees : de data hierboven is geïmporteerd)
+	# de pagina verwijderen
+	if($overzicht OR $detail) {
 		unlink($bestand);
 	}
+	
+	$block[] = implode("\n", $String);
 }
-$block[] = implode("\n", $String);
 
 
 # Laat de resultaten vam de check netjes op het scherm zien.
